split up crypto.cpp/krypto.md and K KRYTPO hook family
